refactor carousel
