+++ pocket_world_application/pom.xml
@@ -51,6 +51,11 @@
             <artifactId>building</artifactId>
             <version>${building.version}</version>
         </dependency>
+        <dependency>
+            <groupId>com.google.guava</groupId>
+            <artifactId>guava</artifactId>
+            <version>18.0</version>
+        </dependency>
     </dependencies>
