+++ pocket_world_application/src/main/java/org/exschool/pocketworld/city/center/service/CityCenterService.java
@@ -2,6 +2,9 @@ package org.exschool.pocketworld.city.center.service;
 
 import org.exschool.pocketworld.city.center.dto.CityCenterDto;
 
+import java.util.List;
+
 public interface CityCenterService {
     CityCenterDto cityCenterInfo();
+    List<String> getBuildingTypesAvailableToBuild();
 }
